requirements.txt
bchlib>=0.7
numpy>=1.16.2
pillow>=5.4.1
Flask==2.2.2
